refactor(architecture): extract research task parser
