+++ tests/unit/test_research.py
@@ -9,14 +9,7 @@ import pytest
 
 import notebooklm._research as research_module
 from notebooklm import NotebookLMClient
-from notebooklm._research import (
-    ResearchAPI,
-    _extract_query_text,
-    _extract_sources_and_summary,
-    _extract_status_code,
-    _extract_task_id,
-    _extract_task_info,
-)
+from notebooklm._research import ResearchAPI
 from notebooklm.rpc import RPCMethod
 
 
@@ -37,34 +30,6 @@ def _build_research_task_payload(
     """Build one POLL_RESEARCH task_info entry for wait/poll tests."""
     sources = [[source_url, source_title, "desc", 1]]
     return [None, [query, 1], 1, [sources, f"{query} summary"], status_code]
-
-
-class TestParseResultType:
-    """Tests for ResearchAPI._parse_result_type static method."""
-
-    def test_int_passthrough(self):
-        assert ResearchAPI._parse_result_type(5) == 5
-
-    def test_known_string_alias(self):
-        assert ResearchAPI._parse_result_type("web") == 1
-        assert ResearchAPI._parse_result_type("drive") == 2
-        assert ResearchAPI._parse_result_type("report") == 5
-
-    def test_case_insensitive(self):
-        assert ResearchAPI._parse_result_type("WEB") == 1
-        assert ResearchAPI._parse_result_type("Drive") == 2
-
-    def test_unknown_string_preserved(self):
-        assert ResearchAPI._parse_result_type("video") == "video"
-
-    def test_none_defaults_to_1(self):
-        assert ResearchAPI._parse_result_type(None) == 1
-
-    def test_float_defaults_to_1(self):
-        assert ResearchAPI._parse_result_type(3.14) == 1
-
-    def test_list_defaults_to_1(self):
-        assert ResearchAPI._parse_result_type([]) == 1
 
 
 class TestBuildImportEntries:
@@ -179,203 +144,6 @@ class TestCitedSourceSelection:
         assert selection.matched_url_source_count == 0
         assert selection.sources == sources
         assert "none of the report URLs matched" in caplog.text
-
-
-class TestExtractLegacyReportChunks:
-    """Tests for _extract_legacy_report_chunks static method."""
-
-    def test_missing_index_6(self):
-        assert ResearchAPI._extract_legacy_report_chunks([None, "t", None, 5, None, None]) == ""
-
-    def test_index_6_not_list(self):
-        assert (
-            ResearchAPI._extract_legacy_report_chunks([None, "t", None, 5, None, None, "str"]) == ""
-        )
-
-    def test_single_chunk(self):
-        assert (
-            ResearchAPI._extract_legacy_report_chunks([None, "t", None, 5, None, None, ["chunk"]])
-            == "chunk"
-        )
-
-    def test_multiple_chunks_joined(self):
-        src = [None, "t", None, 5, None, None, ["a", "b", "c"]]
-        assert ResearchAPI._extract_legacy_report_chunks(src) == "a\n\nb\n\nc"
-
-    def test_filters_non_string_and_empty(self):
-        src = [None, "t", None, 5, None, None, ["real", None, "", 42, "also_real"]]
-        assert ResearchAPI._extract_legacy_report_chunks(src) == "real\n\nalso_real"
-
-    def test_all_empty_returns_empty(self):
-        assert (
-            ResearchAPI._extract_legacy_report_chunks([None, "t", None, 5, None, None, ["", None]])
-            == ""
-        )
-
-
-class TestExtractTaskId:
-    """Tests for ``_extract_task_id`` helper."""
-
-    def test_happy_path(self):
-        assert _extract_task_id(["task_abc", ["info"]]) == "task_abc"
-
-    def test_empty_list_drift_returns_none(self, caplog, monkeypatch):
-        # Post-PR 13.9a default is strict; pin soft mode to keep asserting
-        # the warn-and-return-None contract these helpers expose.
-        monkeypatch.setenv("NOTEBOOKLM_STRICT_DECODE", "0")
-        with (
-            caplog.at_level(logging.WARNING),
-            pytest.warns(DeprecationWarning, match="safe_index soft-mode"),
-        ):
-            assert _extract_task_id([]) is None
-        assert "safe_index drift" in caplog.text
-
-    def test_non_string_id_drift_returns_none(self, caplog):
-        with caplog.at_level(logging.WARNING):
-            assert _extract_task_id([42, ["info"]]) is None
-        assert "task_data[0] is not a string" in caplog.text
-
-    def test_non_list_input_returns_none(self, caplog, monkeypatch):
-        # Soft-mode opt-in: the helper's outer guard returns None without
-        # invoking safe_index; the descent path under strict mode would
-        # otherwise surface UnknownRPCMethodError for the inner safe_index hop.
-        monkeypatch.setenv("NOTEBOOKLM_STRICT_DECODE", "0")
-        with (
-            caplog.at_level(logging.WARNING),
-            pytest.warns(DeprecationWarning, match="safe_index soft-mode"),
-        ):
-            assert _extract_task_id(None) is None
-
-
-class TestExtractTaskInfo:
-    """Tests for ``_extract_task_info`` helper."""
-
-    def test_happy_path(self):
-        info = [None, ["q"], None, [[]], 2]
-        assert _extract_task_info(["task_id", info]) is info
-
-    def test_missing_index_returns_none(self, caplog, monkeypatch):
-        monkeypatch.setenv("NOTEBOOKLM_STRICT_DECODE", "0")
-        with (
-            caplog.at_level(logging.WARNING),
-            pytest.warns(DeprecationWarning, match="safe_index soft-mode"),
-        ):
-            assert _extract_task_info(["only_id"]) is None
-        assert "safe_index drift" in caplog.text
-
-    def test_non_list_value_returns_none(self, caplog):
-        with caplog.at_level(logging.WARNING):
-            assert _extract_task_info(["task_id", "not_a_list"]) is None
-        assert "task_data[1] is not a list" in caplog.text
-
-
-class TestExtractQueryText:
-    """Tests for ``_extract_query_text`` helper."""
-
-    def test_happy_path(self):
-        task_info = [None, ["quantum computing", "extra"], None, [], 1]
-        assert _extract_query_text(task_info) == "quantum computing"
-
-    def test_missing_query_info_returns_none(self, caplog, monkeypatch):
-        monkeypatch.setenv("NOTEBOOKLM_STRICT_DECODE", "0")
-        with (
-            caplog.at_level(logging.WARNING),
-            pytest.warns(DeprecationWarning, match="safe_index soft-mode"),
-        ):
-            # task_info[1] missing entirely
-            assert _extract_query_text([None]) is None
-        assert "safe_index drift" in caplog.text
-
-    def test_non_string_query_returns_none(self, caplog):
-        with caplog.at_level(logging.WARNING):
-            assert _extract_query_text([None, [123], None, [], 1]) is None
-        assert "task_info[1][0] is not a string" in caplog.text
-
-
-class TestExtractStatusCode:
-    """Tests for ``_extract_status_code`` helper."""
-
-    def test_happy_path_in_progress(self):
-        assert _extract_status_code([None, ["q"], None, [], 1]) == 1
-
-    def test_happy_path_completed(self):
-        assert _extract_status_code([None, ["q"], None, [], 2]) == 2
-
-    def test_happy_path_deep_completed(self):
-        assert _extract_status_code([None, ["q"], None, [], 6]) == 6
-
-    def test_missing_index_returns_none(self, caplog, monkeypatch):
-        monkeypatch.setenv("NOTEBOOKLM_STRICT_DECODE", "0")
-        with (
-            caplog.at_level(logging.WARNING),
-            pytest.warns(DeprecationWarning, match="safe_index soft-mode"),
-        ):
-            assert _extract_status_code([None, ["q"], None, []]) is None
-        assert "safe_index drift" in caplog.text
-
-    def test_non_int_returns_none(self, caplog):
-        with caplog.at_level(logging.WARNING):
-            assert _extract_status_code([None, ["q"], None, [], "completed"]) is None
-        assert "task_info[4] is not an int" in caplog.text
-
-    def test_bool_rejected(self, caplog):
-        with caplog.at_level(logging.WARNING):
-            assert _extract_status_code([None, ["q"], None, [], True]) is None
-        assert "task_info[4] is bool" in caplog.text
-
-
-class TestExtractSourcesAndSummary:
-    """Tests for ``_extract_sources_and_summary`` helper."""
-
-    def test_happy_path_with_summary(self):
-        task_info = [
-            None,
-            ["q"],
-            None,
-            [[["https://example.com", "Example"]], "Summary text"],
-            2,
-        ]
-        sources, summary = _extract_sources_and_summary(task_info)
-        assert sources == [["https://example.com", "Example"]]
-        assert summary == "Summary text"
-
-    def test_happy_path_sources_only(self):
-        task_info = [None, ["q"], None, [[["url", "title"]]], 2]
-        sources, summary = _extract_sources_and_summary(task_info)
-        assert sources == [["url", "title"]]
-        assert summary is None
-
-    def test_missing_bundle_returns_empty(self, caplog, monkeypatch):
-        monkeypatch.setenv("NOTEBOOKLM_STRICT_DECODE", "0")
-        with (
-            caplog.at_level(logging.WARNING),
-            pytest.warns(DeprecationWarning, match="safe_index soft-mode"),
-        ):
-            sources, summary = _extract_sources_and_summary([None, ["q"], None])
-        assert sources == []
-        assert summary is None
-        assert "safe_index drift" in caplog.text
-
-    def test_empty_bundle_returns_empty(self):
-        sources, summary = _extract_sources_and_summary([None, ["q"], None, [], 2])
-        assert sources == []
-        assert summary is None
-
-    def test_non_list_bundle_drift(self, caplog):
-        with caplog.at_level(logging.WARNING):
-            sources, summary = _extract_sources_and_summary([None, ["q"], None, "drift", 2])
-        assert sources == []
-        assert summary is None
-        assert "task_info[3] is not a list" in caplog.text
-
-    def test_non_list_sources_slot_drift(self, caplog):
-        with caplog.at_level(logging.WARNING):
-            sources, summary = _extract_sources_and_summary(
-                [None, ["q"], None, ["not_a_list", "Summary"], 2]
-            )
-        assert sources == []
-        assert summary == "Summary"
-        assert "task_info[3][0] is not a list" in caplog.text
 
 
 class TestResearch:
